@@ -1,21 +1,19 @@
 [package]
 authors = ["jD91mZM2 <me@krake.one>"]
 description = "A Nix parser written in Rust"
-edition = "2018"
+edition = "2021"
 license = "MIT"
 name = "rnix"
 readme = "README.md"
 repository = "https://github.com/nix-community/rnix-parser"
-version = "0.9.1"
+version = "0.11.0-dev"
 
 [[bench]]
 harness = false
 name = "all-packages"
 
 [dependencies]
-cbitset = "0.2.0"
-rowan = "0.12.5"
-smol_str = "0.1.17"
+rowan = "0.15.0"
 
 [dev-dependencies]
 criterion = "0.3.0"
